[refactor/U02] main merge 후 폴더 구조 정리
